Updated encdec_paytm.php
I am facing an issue with this file. When (in test mode) I try to pay with debit/credit card the transaction becomes successful but it returns 'Checksum mismatched'. However, it was working fine while paying with net banking. 
Now it is working fine in all the cases.
NOTE: I have done all the tests in 'TEST MODE' only.